+++ mineral/agents/ppo/experience.py
@@ -15,8 +15,7 @@ def transform_op(arr):
 
 
 class ExperienceBuffer(Dataset):
-    def __init__(
-        self, num_envs, horizon_length, batch_size, minibatch_size, obs_space, act_dim, device, cpu_obs_keys):
+    def __init__(self, num_envs, horizon_length, batch_size, minibatch_size, obs_space, act_dim, device, cpu_obs_keys):
         self.device = device
         self.num_envs = num_envs
         self.transitions_per_env = horizon_length
@@ -26,37 +25,29 @@ class ExperienceBuffer(Dataset):
         self.obs_space = obs_space
         self.act_dim = act_dim
         self.storage_dict = {
-            'rewards': torch.zeros(
-                (self.transitions_per_env, self.num_envs, 1),
-                dtype=torch.float32, device=self.device),
-            'values': torch.zeros(
-                (self.transitions_per_env, self.num_envs, 1),
-                dtype=torch.float32, device=self.device),
-            'neglogp': torch.zeros(
-                (self.transitions_per_env, self.num_envs),
-                dtype=torch.float32, device=self.device),
-            'dones': torch.zeros(
-                (self.transitions_per_env, self.num_envs),
-                dtype=torch.uint8, device=self.device),
+            'rewards': torch.zeros((self.transitions_per_env, self.num_envs, 1), dtype=torch.float32, device=self.device),
+            'values': torch.zeros((self.transitions_per_env, self.num_envs, 1), dtype=torch.float32, device=self.device),
+            'neglogp': torch.zeros((self.transitions_per_env, self.num_envs), dtype=torch.float32, device=self.device),
+            'dones': torch.zeros((self.transitions_per_env, self.num_envs), dtype=torch.uint8, device=self.device),
             'actions': torch.zeros(
-                (self.transitions_per_env, self.num_envs, self.act_dim),
-                dtype=torch.float32, device=self.device),
+                (self.transitions_per_env, self.num_envs, self.act_dim), dtype=torch.float32, device=self.device
+            ),
             'mus': torch.zeros(
-                (self.transitions_per_env, self.num_envs, self.act_dim),
-                dtype=torch.float32, device=self.device),
+                (self.transitions_per_env, self.num_envs, self.act_dim), dtype=torch.float32, device=self.device
+            ),
             'sigmas': torch.zeros(
-                (self.transitions_per_env, self.num_envs, self.act_dim),
-                dtype=torch.float32, device=self.device),
-            'returns': torch.zeros(
-                (self.transitions_per_env, self.num_envs, 1),
-                dtype=torch.float32, device=self.device),
+                (self.transitions_per_env, self.num_envs, self.act_dim), dtype=torch.float32, device=self.device
+            ),
+            'returns': torch.zeros((self.transitions_per_env, self.num_envs, 1), dtype=torch.float32, device=self.device),
         }
 
         obs_dict = {}
         for k, v in self.obs_space.items():
             buffer = torch.zeros(
                 (self.transitions_per_env, self.num_envs, *v),
-                dtype=torch.float32, device=self.device if not re.match(self.cpu_obs_keys, k) else 'cpu')
+                dtype=torch.float32,
+                device=self.device if not re.match(self.cpu_obs_keys, k) else 'cpu',
+            )
             obs_dict[k] = buffer
         self.storage_dict['obses'] = obs_dict
 
@@ -78,9 +69,16 @@ class ExperienceBuffer(Dataset):
                 input_dict[k] = v_dict
             else:
                 input_dict[k] = v[start:end]
-        return input_dict['values'], input_dict['neglogp'], input_dict['advantages'], \
-            input_dict['mus'], input_dict['sigmas'], input_dict['returns'], input_dict['actions'], \
-            input_dict['obses']
+        return (
+            input_dict['values'],
+            input_dict['neglogp'],
+            input_dict['advantages'],
+            input_dict['mus'],
+            input_dict['sigmas'],
+            input_dict['returns'],
+            input_dict['actions'],
+            input_dict['obses'],
+        )
 
     def update_mu_sigma(self, mu, sigma):
         start = self.last_range[0]
@@ -91,9 +89,9 @@ class ExperienceBuffer(Dataset):
     def update_data(self, name, index, val):
         if type(val) is dict:
             for k, v in val.items():
-                self.storage_dict[name][k][index,:] = v
+                self.storage_dict[name][k][index, :] = v
         else:
-            self.storage_dict[name][index,:] = val
+            self.storage_dict[name][index, :] = val
 
     def compute_return(self, last_values, gamma, tau):
         last_gae_lam = 0
@@ -105,8 +103,7 @@ class ExperienceBuffer(Dataset):
                 next_values = self.storage_dict['values'][t + 1]
             next_nonterminal = 1.0 - self.storage_dict['dones'].float()[t]
             next_nonterminal = next_nonterminal.unsqueeze(1)
-            delta = self.storage_dict['rewards'][t] + \
-                gamma * next_values * next_nonterminal - self.storage_dict['values'][t]
+            delta = self.storage_dict['rewards'][t] + gamma * next_values * next_nonterminal - self.storage_dict['values'][t]
             mb_advs[t] = last_gae_lam = delta + gamma * tau * next_nonterminal * last_gae_lam
             self.storage_dict['returns'][t, :] = mb_advs[t] + self.storage_dict['values'][t]
 
@@ -120,6 +117,5 @@ class ExperienceBuffer(Dataset):
             else:
                 self.data_dict[k] = transform_op(v)
         advantages = self.data_dict['returns'] - self.data_dict['values']
-        self.data_dict['advantages'] = (
-            (advantages - advantages.mean()) / (advantages.std() + 1e-8)).squeeze(1)
+        self.data_dict['advantages'] = ((advantages - advantages.mean()) / (advantages.std() + 1e-8)).squeeze(1)
         return self.data_dict
