+++ mineral/cfgs/agent/CartpolePPO.yaml
@@ -1,12 +1,16 @@
 seed: ${..seed}
 algo: PPO
 
+save_video_every: 0
+save_video_consecutive: 0
+
 network:
   mlp:
     units: [32, 32]
   separate_value_mlp: True
   act_type: ELU
   norm_type: null
+  actor_input_ind_std: True
 
 ppo:
   name: ${resolve_default:Cartpole,${...experiment}}
@@ -26,10 +30,6 @@ ppo:
   adam_eps: 1e-5
   lr_schedule: kl  # 'fixed' | 'linear' | 'kl' | 'cos'
   kl_threshold: 0.008
-  save_best_after: 500
-  save_frequency: 100
-  save_video_every: 0
-  save_video_consecutive: 0
   grad_norm: 1.0
   entropy_coef: 0.0
   truncate_grads: True
@@ -42,3 +42,5 @@ ppo:
   bounds_loss_coef: 0.0001
   bounds_type: bound
   max_agent_steps: 1200000
+  save_best_after: 500
+  save_frequency: 100
