+++ mineral/cfgs/config.yaml
@@ -1,9 +1,29 @@
 
+# set default task and default training config based on task
+defaults:
+  - task: Cartpole
+  - agent: ${task}PPO
+  - _self_
+
+  # - hydra/job_logging: disabled
+  # - override hydra/job_logging: colorlog
+  # - override hydra/hydra_logging: colorlog
+
+# set the directory where the output files get saved
+hydra:
+  searchpath:
+    - pkg://isaacgymenvs/cfg
+  output_subdir: null
+  run:
+    dir: .
+  job_logging:
+    formatters:
+      simple:
+        format: "[%(levelname)s: %(module)s:%(lineno)d %(asctime)s] %(message)s"
+
 # Task name - used to pick the class to load
 task_name: ${task.name}
-# experiment name. defaults to name of training config
 experiment: ''
-
 # if set to positive integer, overrides the default number of environments
 num_envs: ''
 
@@ -12,32 +32,7 @@ seed: 42
 # set to True for deterministic performance
 torch_deterministic: False
 
-## Device config
-#  'physx' or 'flex'
-physics_engine: 'physx'
-# whether to use cpu or gpu pipeline
-pipeline: 'gpu'
-
-device_id: 0  # 'cuda:?', -1 for 'cpu'
-# device for running physics simulation
-sim_device: ???
-# device to run RL
-rl_device: ???
-graphics_device_id: 0
-
-## PhysX arguments
-num_threads: 4 # Number of worker threads per scene used by PhysX - for CPU PhysX only.
-solver_type: 1 # 0: pgs, 1: tgs
-num_subscenes: 4 # Splits the simulation into N physics scenes and runs each one in a separate thread
-
-# RL Arguments
-# test - if set, run policy in inference mode (requires setting checkpoint to load)
-test: False
-# used to set checkpoint path
-checkpoint: ''
-# set to True to use multi-gpu horovod training
-multi_gpu: False
-
+## Logging
 # logdir: 'debug'
 logdir: 'workdir/${task.name}'
 
@@ -48,27 +43,29 @@ wandb:
   resume: allow
   # save_code: True
 
+## PhysX arguments
+physics_engine: 'physx'  # physx | flex
+pipeline: 'gpu'  # gpu | cpu
+num_threads: 4  # Number of worker threads per scene used by PhysX - for CPU PhysX only.
+solver_type: 1  # 0: pgs, 1: tgs
+num_subscenes: 4  # Splits the simulation into N physics scenes and runs each one in a separate thread
+
+## Device config
+device_id: 0  # 'cuda:?', -1 for 'cpu'
+# device for running physics simulation
+sim_device: ???
+# device to run RL
+rl_device: ???
+graphics_device_id: 0
 # disables rendering
 headless: True
-
 # whether to call env.render(mode='rgb_array')
 env_render: False
 
-# set default task and default training config based on task
-defaults:
-  - task: Cartpole
-  - train: ${task}PPO
-
-  - hydra/job_logging: disabled
-  # - override hydra/job_logging: colorlog
-  # - override hydra/hydra_logging: colorlog
-
-# set the directory where the output files get saved
-hydra:
-  output_subdir: null
-  run:
-    dir: .
-  job_logging:
-    formatters:
-      simple:
-        format: "[%(levelname)s: %(module)s:%(lineno)d %(asctime)s] %(message)s"
+## RL Arguments
+# test - if set, run policy in inference mode (requires setting checkpoint to load)
+test: False
+# used to set checkpoint path
+checkpoint: ''
+# set to True to use multi-gpu horovod training
+multi_gpu: False
